# api.yaml
swagger: '2.0'
info:
  title: Reshigo API
  version: "1.0"
# the domain of the service
host: server.reshigo.ru
...
schemes:
  - http
# will be prefixed to all paths
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
    description: Http Basic Authentication. Works over `HTTP`
paths:
  /password:
    get:
      description: Request 4-digit code to change password
      tags:
        - Consumer
        - Solver
      parameters:
        - name: name
          in: query
          required: false
          type: string
        - name: phone
          in: query
          required: false
          type: string
      responses:
        200:
          description: OK
        403:
          description: User is disabled, contact support.
        404:
          description: User not found
        500:
          description: Some problem with phone number. Contact support.
    put:
      description: update password
      tags:
       - Consumer
       - Solver
      parameters:
        - name: recovery
          in: body
          description: password field must be valid
          required: true
          schema:
            $ref: '#/definitions/PassRecovery'
      responses:
        200:
          description: OK
        400:
          description: object not valid
        403:
          description: user disabled, contact support
        404:
          description: user not found
  /password/resend:
    get:
      description: resend password recovery code, if sms was not received
      tags:
        - Consumer
        - Solver
      parameters:
        - name: name
          in: query
          required: false
          type: string
        - name: phone
          in: query
          required: false
          type: string
      responses:
        200:
          description: OK
        403:
          description: User is disabled, contact support.
        404:
          description: User not found
        500:
          description: Some problem with phone number. Contact support.
  /code/validate:
    post:
      description: validate 4-digit code. if OK, should ask for new password and PUT it to /password. 3 attempts max.
      tags:
       - Consumer
       - Solver
      parameters:
        - name: recovery
          in: body
          description: password field can be null
          required: true
          schema:
            $ref: '#/definitions/PassRecovery'
      responses:
        200:
          description: OK
        400:
          description: code or object not valid
        403:
          description: user disabled, contact support
        404:
          description: user not found
...
definitions:
...
  PassRecovery:
    type: object
    properties:
      name:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+
      phone:
        type: string
      code:
        type: integer
        format: int64
      password:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+
...